whatsthedata_saas/requirements.txt
# WhatTheData SaaS Requirements
# ================================

# Framework principal
streamlit==1.31.0
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Base de données
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentification et sécurité
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7

# APIs externes
stripe==8.0.0
requests==2.31.0
httpx==0.26.0

# Configuration
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Data processing
pandas==2.1.4
numpy==1.26.3
openpyxl==3.1.2

# Visualisation
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.1

# LinkedIn API
linkedin-api==2.2.0  # Si disponible, sinon requests

# Facebook API
facebook-sdk==3.1.0

# Utilitaires
python-dateutil==2.8.2
pytz==2023.4
schedule==1.2.1
loguru==0.7.2

# OAuth et JWT
authlib==1.3.0
oauthlib==3.2.2

# Monitoring et logs
structlog==23.2.0

# Development tools
black==24.1.1
flake8==7.0.0
pytest==7.4.4
pytest-asyncio==0.23.3

# Cache (optionnel)
redis==5.0.1

# File processing
PyPDF2==3.0.1
python-magic==0.4.27